chore: update rusqlite from 0.37 to 0.40.2
Largest change comes from removal of u64 and usize serialization
in rusqlite 0.38.0.
It is possible to enable it back with `fallible_uint` feature
as suggested in https://github.com/rusqlite/rusqlite/pull/1732,
but there is a good reason for this change
as it may result in unexpected failures when
trying to store u64 integers that don't fit in i64,
so I don't enable it back.
With this change conversions between i64 and u64 are explicit.

Using less `usize` is generally good as its size
differs on 32-bit platforms that are rarely tested
but still supported and we have changes that reduce its usage
such as 0defa117a0a302985c61fe211d1a6a948a06d4f6,
014d2ace76790a8484609a4a27fce53ba2fdaba3 etc.

+++ src/reaction.rs
@@ -84,7 +84,7 @@ pub struct ReactionFrequency {
     pub reaction: Reaction,
 
     /// Number of contacts that reacted with this emoji.
-    pub count: usize,
+    pub count: u32,
 
     /// True if `ContactId::SELF` is among the contacts that reacted with this emoji.
     pub is_from_self: bool,
@@ -420,7 +420,7 @@ pub(crate) async fn apply_pending_reactions(
 /// sorted in descending order of frequency.
 fn calc_frequencies(by_contact: &BTreeMap<ContactId, Reaction>) -> Vec<ReactionFrequency> {
     let mut self_reaction = Reaction::new("");
-    let mut counts: BTreeMap<&str, usize> = BTreeMap::new();
+    let mut counts: BTreeMap<&str, u32> = BTreeMap::new();
     for (contact_id, reaction) in by_contact {
         let count = counts.entry(reaction.as_str()).or_insert(0);
         *count = count.saturating_add(1);
